Project reorganize
Moved lots of things out of main and into their own modules.
