+++ ollyfc/src/lib.rs
@@ -1,5 +1,3 @@
 #![no_std]
-pub mod flight_control;
-pub mod flight_logger;
-pub mod sbus;
+
 pub mod w25q;
